Adds shift deletion functionality
Implements functionality to delete shifts via alert dialog.

This change introduces the `@radix-ui/react-alert-dialog` component to confirm shift deletions.
It also adds a new server action `deleteShift` to handle the deletion logic and revalidate the shifts route.
Finally, it refactors the shifts table to display the alert dialog when deleting a shift.