Adds shift creation and improves shift actions (#2)
* Adds shift creation functionality.

Implements the ability to add new shifts to the application.

This includes:
- Adding necessary UI components (dialog, form, datepicker)
- Integrating with react-hook-form for form management and validation
- Implementing a server action to handle shift creation and database interaction
- Adding validation to the shift creation process.
- Adding radix ui packages.

Also adds .idea to .gitignore.

* Adds time inputs to shift form

Updates the shift form to include time inputs for shift start and end times.

Combines the date and time inputs into a single Date object for submission.

Adds validation to ensure the end date and time is after the start date and time.

* Improves shift action dropdown management

Refactors the shift action dropdown into a separate `ActionsCell` component to manage its open state internally. This fixes an issue where the dropdown state was not being properly maintained, especially after deleting a shift. It also adds a callback to the `DeleteShiftDialog` to close the parent dropdown upon successful deletion.